split out kimchi_types into file
